+++ web/src/components/organisms/Asset/Asset/hooks.ts
@@ -2,7 +2,7 @@ import { useCallback, useEffect, useMemo, useState } from "react";
 
 import Notification from "@reearth-cms/components/atoms/Notification";
 import { Asset, PreviewType } from "@reearth-cms/components/molecules/Asset/asset.type";
-import { viewerRef } from "@reearth-cms/components/molecules/Asset/Asset/AssetBody/index";
+import { viewerRef } from "@reearth-cms/components/molecules/Asset/Asset/AssetBody/Asset";
 import {
   Asset as GQLAsset,
   PreviewType as GQLPreviewType,
